feat(eval): score repeated runtime failures
